+++ linen-repository/src/main/scala/x7c1/linen/database/mixin/ChannelsToAttachRecord.scala
@@ -2,8 +2,7 @@ package x7c1.linen.database.mixin
 
 import android.database.Cursor
 import x7c1.linen.database.struct.ChannelRecord
-import x7c1.wheat.macros.database.TypedCursor
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.{Query, TypedCursor}
 import x7c1.wheat.modern.database.selector.presets.CanTraverseRecord
 import x7c1.wheat.modern.database.selector.{IdEndo, Identifiable, RecordReifiable}
