+++ linen-repository/src/main/scala/x7c1/linen/database/mixin/LoaderScheduleWithKind.scala
@@ -3,8 +3,7 @@ package x7c1.linen.database.mixin
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
 import x7c1.linen.database.struct.{HasAccountId, HasLoaderScheduleId, LoaderScheduleKindRecord, LoaderScheduleRecord}
-import x7c1.wheat.macros.database.TypedCursor
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.{Query, TypedCursor}
 import x7c1.wheat.modern.database.selector.presets.{CanFindRecord, CanTraverseRecord, CanTraverseRecordByQuery, Find, FindBy, TraverseAll, TraverseOn}
 import x7c1.wheat.modern.database.selector.{RecordReifiable, SelectorProvidable}
