+++ linen-repository/src/main/scala/x7c1/linen/database/mixin/MyChannelRecord.scala
@@ -2,8 +2,7 @@ package x7c1.linen.database.mixin
 
 import android.database.Cursor
 import x7c1.linen.database.struct.{ChannelRecord, ChannelStatusRecord, HasAccountId}
-import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.{Query, TypedCursor, TypedFields}
 import x7c1.wheat.modern.database.selector.RecordReifiable
 import x7c1.wheat.modern.database.selector.presets.CanTraverseRecord
