+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/preset/AllPresetChannelsAccessor.scala
@@ -1,6 +1,6 @@
 package x7c1.linen.repository.channel.preset
 
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.Query
 
 object AllPresetChannelsAccessor
   extends PresetChannelAccessorFactory(AllPresetChannelsQuery)
