+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/preset/PresetChannelAccessorFactory.scala
@@ -3,7 +3,7 @@ package x7c1.linen.repository.channel.preset
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.repository.account.PresetAccount
 import x7c1.linen.repository.preset.{NoPresetAccount, PresetRecordError, UnexpectedException}
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.Query
 
 class PresetChannelAccessorFactory(queryFactory: PresetChannelQueryFactory){
   def create(
