+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/preset/PresetChannelsAccessor.scala
@@ -1,8 +1,7 @@
 package x7c1.linen.repository.channel.preset
 
 import x7c1.linen.database.control.DatabaseHelper
-import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.{Query, TypedCursor, TypedFields}
 
 trait PresetChannelsAccessor {
   def clientAccountId: Long
