+++ wheat-modern/src/main/scala/x7c1/wheat/modern/database/selector/presets/CanCollectRecord.scala
@@ -1,7 +1,8 @@
 package x7c1.wheat.modern.database.selector.presets
 
 import android.database.Cursor
-import x7c1.wheat.modern.database.{HasTable, Query}
+import x7c1.wheat.macros.database.Query
+import x7c1.wheat.modern.database.HasTable
 import x7c1.wheat.modern.database.selector.{CanIdentify, CursorConverter, CursorReadable, CursorReifiable}
 
 import scala.language.{higherKinds, reflectiveCalls}
