+++ wheat-modern/src/main/scala/x7c1/wheat/modern/database/selector/presets/CanTraverseEntity.scala
@@ -1,7 +1,7 @@
 package x7c1.wheat.modern.database.selector.presets
 
 import android.database.{Cursor, SQLException}
-import x7c1.wheat.modern.database.Query
+import x7c1.wheat.macros.database.Query
 import x7c1.wheat.modern.database.selector.{CanIdentify, CursorReadable, CursorReifiable}
 
 import scala.language.{higherKinds, reflectiveCalls}
